[REFACTOR] - 다운로드 버튼 컴포넌트화
